Change distance sensor to return actual distance instead of normalized distance

+++ myPacManProject/src/pacman/entries/pacman/newNeuralPacMan/neurons/sensors/Bias.java
@@ -14,7 +14,7 @@ public class Bias extends Sensor {
 	}
 
 	@Override
-	protected float getNormalizedSensorValue(int pacManIndex,
+	protected double getNormalizedSensorValue(int pacManIndex,
 		int[] validIndexes, Game game) {
 		return 1;
 	}
